Start refactor of user_auth and user.service to properly separate concerns.
Separate HTTP and API functionality, user data, and auth from auth.service into: userApi.service, user.store, and auth.service.
