+++ frontend/angular-fleets/src/app/components/user/user.component.html
@@ -48,7 +48,18 @@
             </ng-template>
             Account Tabs
           </button>
+
+          <!-- DELETE LISTING BUTTON -->
+          <!-- <button mat-button class="action-buttons" [disabled]=".selection?.hasValue()">
+
+
+          </button>-->
         </div>
+
+
+
+
+
         <div class="account-window-body">
 
           <!-- PROFILE TAB -->
